feat: 모델타입 Pick 영역 API 연결
