--- frontend/src/pages/ModelTypePage/Pick/HMGData/constants.js
@@ -1,13 +1,0 @@
-export const HMG_DATA = {
-  INFO: '배기량과 평균연비입니다.',
-  AND: `와`,
-  OF: '의',
-  DISPLACEMENT: '배기량',
-  FUELEFFICIENCY: '평균연비',
-  DISPLACEMENT_UNIT: 'cc',
-  FUELEFFICIENCY_UNIT: 'km/s',
-};
-
-export const HMG_TAG = {
-  TYPE: 'tag20',
-};
