+++ frontend/src/pages/ModelTypePage/Pick/HMGData/index.jsx
@@ -1,15 +1,16 @@
 import * as S from './style';
 import { useData } from '../../../../utils/Context';
-import { HMG_DATA, HMG_TAG } from './constants';
+import { HMG_DATA, HMG_TAG } from '../../constants';
 import HMGTag from '../../../../components/HMGTag';
 import HMGUnit from './HMGUnit';
 
-// ! API 연결 후 수정 필요
+// !FIXME Performance API 연결
 const cc = 2199;
 const km = 12;
 
 function HMGData() {
   const { modelType } = useData();
+
   const tagProps = { type: HMG_TAG.TYPE };
   const displacementProps = {
     title: HMG_DATA.DISPLACEMENT,
@@ -25,9 +26,9 @@ function HMGData() {
       <S.Info>
         <HMGTag {...tagProps} />
         <S.InfoText>
-          <S.InfoHighlight>디젤2.2</S.InfoHighlight>
+          <S.InfoHighlight>{modelType.powerTrainOption.name}</S.InfoHighlight>
           {HMG_DATA.AND}&nbsp;
-          <S.InfoHighlight>2WD</S.InfoHighlight>
+          <S.InfoHighlight>{modelType.wheelDriveOption.name}</S.InfoHighlight>
           {HMG_DATA.OF}
         </S.InfoText>
         <S.InfoText>{HMG_DATA.INFO}</S.InfoText>
