+++ frontend/src/pages/ModelTypePage/constants.js
@@ -11,3 +11,17 @@ export const PICK_CARD = {
   BODY_TYPE: '바디타입',
   WHEEL_DRIVE: '구동방식',
 };
+
+export const HMG_DATA = {
+  INFO: '배기량과 평균연비입니다.',
+  AND: `와`,
+  OF: '의',
+  DISPLACEMENT: '배기량',
+  FUELEFFICIENCY: '평균연비',
+  DISPLACEMENT_UNIT: 'cc',
+  FUELEFFICIENCY_UNIT: 'km/s',
+};
+
+export const HMG_TAG = {
+  TYPE: 'tag20',
+};
